ion: isolate msm-specific ion extensions

+++ arch/arm/mach-msm/board-8064.c
@@ -454,7 +454,7 @@ static void __init reserve_ion_memory(void)
 		if (heap->extra_data) {
 			int fixed_position = NOT_FIXED;
 
-			switch (heap->type) {
+			switch ((int) heap->type) {
 			case ION_HEAP_TYPE_CP:
 				fixed_position = ((struct ion_cp_heap_pdata *)
 					heap->extra_data)->fixed_position;
